refactor(phase1): implement App\Http\Action constants; replace Router switch-case and login checks to use constants; remove old Actions.php
